Twig extension refactoring
