--- src/Sylius/Bundle/AdminBundle/Resources/config/services/templating.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Paweł Jędrzejewski
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <services>
-        <defaults public="true" />
-
-        <service id="sylius.templating.helper.variant_price_helper" class="Sylius\Bundle\AdminBundle\Templating\Helper\VariantPriceHelper" lazy="true">
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.money_formatter" />
-            <tag name="templating.helper" alias="sylius_variant_price" />
-        </service>
-        <service id="sylius.templating.helper.available_locales_helper" class="Sylius\Bundle\AdminBundle\Templating\Helper\AvailableLocaleHelper" lazy="true">
-            <argument type="service" id="sylius.translation_locale_provider.admin" />
-            <tag name="templating.helper" alias="sylius_available_locales" />
-        </service>
-    </services>
-</container>
